Rename API package
Close #83
