+++ app/src/main/java/lt/vilnius/tvarkau/ChooseReportTypeActivity.java
@@ -15,12 +15,12 @@ import autodagger.AutoComponent;
 import autodagger.AutoInjector;
 import butterknife.BindView;
 import butterknife.ButterKnife;
-import lt.vilnius.tvarkau.api.ApiMethod;
-import lt.vilnius.tvarkau.api.ApiRequest;
-import lt.vilnius.tvarkau.api.ApiResponse;
-import lt.vilnius.tvarkau.api.GetProblemTypesParams;
-import lt.vilnius.tvarkau.api.LegacyApiModule;
-import lt.vilnius.tvarkau.api.LegacyApiService;
+import lt.vilnius.tvarkau.backend.ApiMethod;
+import lt.vilnius.tvarkau.backend.ApiRequest;
+import lt.vilnius.tvarkau.backend.ApiResponse;
+import lt.vilnius.tvarkau.backend.GetProblemTypesParams;
+import lt.vilnius.tvarkau.backend.LegacyApiModule;
+import lt.vilnius.tvarkau.backend.LegacyApiService;
 import lt.vilnius.tvarkau.views.adapters.ReportTypesListAdapter;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.functions.Action1;
