+++ app/src/main/java/lt/vilnius/tvarkau/NewProblemActivity.java
@@ -2,7 +2,6 @@ package lt.vilnius.tvarkau;
 
 import android.app.ProgressDialog;
 import android.content.ClipData;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.location.Address;
@@ -19,7 +18,6 @@ import android.support.v4.content.ContextCompat;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -37,7 +35,6 @@ import com.google.android.gms.maps.model.LatLng;
 import com.viewpagerindicator.CirclePageIndicator;
 
 import org.greenrobot.eventbus.EventBus;
-import org.threeten.bp.LocalDate;
 import org.threeten.bp.LocalDateTime;
 import org.threeten.bp.ZoneOffset;
 
@@ -55,16 +52,15 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import icepick.State;
-import lt.vilnius.tvarkau.api.ApiMethod;
-import lt.vilnius.tvarkau.api.ApiRequest;
-import lt.vilnius.tvarkau.api.ApiResponse;
-import lt.vilnius.tvarkau.api.GetNewProblemParams;
-import lt.vilnius.tvarkau.api.LegacyApiModule;
-import lt.vilnius.tvarkau.api.LegacyApiService;
+import lt.vilnius.tvarkau.backend.ApiMethod;
+import lt.vilnius.tvarkau.backend.ApiRequest;
+import lt.vilnius.tvarkau.backend.ApiResponse;
+import lt.vilnius.tvarkau.backend.GetNewProblemParams;
+import lt.vilnius.tvarkau.backend.LegacyApiModule;
+import lt.vilnius.tvarkau.backend.LegacyApiService;
 import lt.vilnius.tvarkau.entity.Profile;
 import lt.vilnius.tvarkau.events_listeners.NewProblemAddedEvent;
 import lt.vilnius.tvarkau.utils.FormatUtils;
-import lt.vilnius.tvarkau.utils.GlobalConsts;
 import lt.vilnius.tvarkau.utils.ImageUtils;
 import lt.vilnius.tvarkau.utils.KeyboardUtils;
 import lt.vilnius.tvarkau.utils.PermissionUtils;
