+++ app/src/main/java/lt/vilnius/tvarkau/backend/ApiMethod.java
@@ -1,4 +1,4 @@
-package lt.vilnius.tvarkau.api;
+package lt.vilnius.tvarkau.backend;
 
 public enum ApiMethod {
