+++ app/src/main/java/lt/vilnius/tvarkau/backend/GetVilniusSignParams.java
@@ -1,4 +1,4 @@
-package lt.vilnius.tvarkau.api;
+package lt.vilnius.tvarkau.backend;
 
 public class GetVilniusSignParams {
