+++ app/src/main/java/lt/vilnius/tvarkau/fragments/ProblemDetailFragment.java
@@ -33,12 +33,12 @@ import lt.vilnius.tvarkau.ProblemDetailActivity;
 import lt.vilnius.tvarkau.ProblemsListActivity;
 import lt.vilnius.tvarkau.ProblemsMapActivity;
 import lt.vilnius.tvarkau.R;
-import lt.vilnius.tvarkau.api.ApiMethod;
-import lt.vilnius.tvarkau.api.ApiRequest;
-import lt.vilnius.tvarkau.api.ApiResponse;
-import lt.vilnius.tvarkau.api.GetProblemParams;
-import lt.vilnius.tvarkau.api.LegacyApiModule;
-import lt.vilnius.tvarkau.api.LegacyApiService;
+import lt.vilnius.tvarkau.backend.ApiMethod;
+import lt.vilnius.tvarkau.backend.ApiRequest;
+import lt.vilnius.tvarkau.backend.ApiResponse;
+import lt.vilnius.tvarkau.backend.GetProblemParams;
+import lt.vilnius.tvarkau.backend.LegacyApiModule;
+import lt.vilnius.tvarkau.backend.LegacyApiService;
 import lt.vilnius.tvarkau.entity.Problem;
 import lt.vilnius.tvarkau.utils.FormatUtils;
 import lt.vilnius.tvarkau.utils.GlobalConsts;
