+++ app/src/main/java/lt/vilnius/tvarkau/fragments/ProblemsListFragment.java
@@ -12,10 +12,8 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.View.OnClickListener;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
-import android.widget.TextView;
 
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
@@ -37,13 +35,13 @@ import lt.vilnius.tvarkau.LogApp;
 import lt.vilnius.tvarkau.NewProblemActivity;
 import lt.vilnius.tvarkau.R;
 import lt.vilnius.tvarkau.SharedPreferencesModule;
-import lt.vilnius.tvarkau.api.ApiMethod;
-import lt.vilnius.tvarkau.api.ApiRequest;
-import lt.vilnius.tvarkau.api.ApiResponse;
-import lt.vilnius.tvarkau.api.GetProblemParams;
-import lt.vilnius.tvarkau.api.GetProblemsParams;
-import lt.vilnius.tvarkau.api.LegacyApiModule;
-import lt.vilnius.tvarkau.api.LegacyApiService;
+import lt.vilnius.tvarkau.backend.ApiMethod;
+import lt.vilnius.tvarkau.backend.ApiRequest;
+import lt.vilnius.tvarkau.backend.ApiResponse;
+import lt.vilnius.tvarkau.backend.GetProblemParams;
+import lt.vilnius.tvarkau.backend.GetProblemsParams;
+import lt.vilnius.tvarkau.backend.LegacyApiModule;
+import lt.vilnius.tvarkau.backend.LegacyApiService;
 import lt.vilnius.tvarkau.entity.Problem;
 import lt.vilnius.tvarkau.events_listeners.EndlessRecyclerViewScrollListener;
 import lt.vilnius.tvarkau.events_listeners.NewProblemAddedEvent;
