+++ app/src/main/java/lt/vilnius/tvarkau/fragments/ReportImportDialogFragment.java
@@ -36,13 +36,13 @@ import lt.vilnius.tvarkau.LogApp;
 import lt.vilnius.tvarkau.NewProblemActivity;
 import lt.vilnius.tvarkau.R;
 import lt.vilnius.tvarkau.SharedPreferencesModule;
-import lt.vilnius.tvarkau.api.ApiMethod;
-import lt.vilnius.tvarkau.api.ApiRequest;
-import lt.vilnius.tvarkau.api.ApiResponse;
-import lt.vilnius.tvarkau.api.GetProblemsParams;
-import lt.vilnius.tvarkau.api.GetVilniusSignParams;
-import lt.vilnius.tvarkau.api.LegacyApiModule;
-import lt.vilnius.tvarkau.api.LegacyApiService;
+import lt.vilnius.tvarkau.backend.ApiMethod;
+import lt.vilnius.tvarkau.backend.ApiRequest;
+import lt.vilnius.tvarkau.backend.ApiResponse;
+import lt.vilnius.tvarkau.backend.GetProblemsParams;
+import lt.vilnius.tvarkau.backend.GetVilniusSignParams;
+import lt.vilnius.tvarkau.backend.LegacyApiModule;
+import lt.vilnius.tvarkau.backend.LegacyApiService;
 import lt.vilnius.tvarkau.entity.LoginResponse;
 import lt.vilnius.tvarkau.entity.Problem;
 import lt.vilnius.tvarkau.events_listeners.NewProblemAddedEvent;
